code review comments addressed
