+++ tests/integration/test_row_checks_geo.py
@@ -25,7 +25,29 @@ from databricks.labs.dqx.geo.check_funcs import (
     is_point,
     is_polygon,
     is_ogc_valid,
+    is_geo_contains,
+    is_geo_covers,
+    is_geo_intersects,
+    is_geo_touches,
+    is_geo_within,
 )
+
+_POINT_INSIDE = "POINT(4.90 52.37)"
+_POINT_EDGE = "POINT(4.73 52.28)"
+_POINT_OUTSIDE = "POINT(4.48 51.92)"
+_REF_POLYGON = "POLYGON((4.73 52.28, 5.05 52.28, 5.05 52.43, 4.73 52.43, 4.73 52.28))"
+_OVERLAPPING_POLYGON = "POLYGON((4.95 52.25, 5.30 52.25, 5.30 52.40, 4.95 52.40, 4.95 52.25))"
+_GEO_SCHEMA = "geom: string"
+_PRECISE_SCHEMA = "geom: string, geom_outside_reference_geometry: string"
+_APPROXIMATE_SCHEMA = "geom: string, geom_outside_reference_geometry_approximate: string"
+
+
+def _precise_violation(value: str) -> str:
+    return f"value `{value}` in column `geom` has no relationship with the reference geometry"
+
+
+def _approximate_violation(value: str) -> str:
+    return f"value `{value}` in column `geom` is approximately outside the reference geometry"
 
 
 def test_is_geometry(skip_if_runtime_not_geo_compatible, spark):
@@ -764,3 +786,320 @@ def test_is_num_points_not_less_than(skip_if_runtime_not_geo_compatible, spark):
     )
 
     assertDataFrameEqual(actual, expected)
+
+
+def test_is_geo_contains_interior_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point is fully contained — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_contains("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_contains_boundary_point_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Boundary point is not contained (st_contains excludes boundary) — violation."""
+    test_df = spark.createDataFrame([[_POINT_EDGE], [None]], _GEO_SCHEMA)
+    condition = is_geo_contains("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_EDGE, _precise_violation(_POINT_EDGE)], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_contains_exterior_point_wkt_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point is not contained — violation (WKT input)."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_contains("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _precise_violation(_POINT_OUTSIDE)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_precise_interior_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point is covered — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_precise_boundary_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Boundary point is covered (st_covers includes boundary, unlike st_contains) — no violation."""
+    test_df = spark.createDataFrame([[_POINT_EDGE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_EDGE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_precise_exterior_point_wkt_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point is not covered — violation (WKT input)."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _precise_violation(_POINT_OUTSIDE)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_interior_point_resolution_5_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point covered by H3 cells at resolution 5 — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_interior_point_resolution_7_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point covered by H3 cells at resolution 7 — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_exterior_point_resolution_5_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point not covered by H3 cells at resolution 5 — violation."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _approximate_violation(_POINT_OUTSIDE)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_exterior_point_resolution_7_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point not covered by H3 cells at resolution 7 — violation."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _approximate_violation(_POINT_OUTSIDE)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_overlapping_polygon_resolution_5_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Overlapping polygon extends beyond reference — not all H3 cells covered at resolution 5 — violation."""
+    test_df = spark.createDataFrame([[_OVERLAPPING_POLYGON], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_OVERLAPPING_POLYGON, _approximate_violation(_OVERLAPPING_POLYGON)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_approximate_overlapping_polygon_resolution_7_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Overlapping polygon extends beyond reference — not all H3 cells covered at resolution 7 — violation."""
+    test_df = spark.createDataFrame([[_OVERLAPPING_POLYGON], [None]], _GEO_SCHEMA)
+    condition = is_geo_covers("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_OVERLAPPING_POLYGON, _approximate_violation(_OVERLAPPING_POLYGON)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_covers_precise_flags_near_boundary_point_approximate_does_not(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Point just outside the polygon boundary (~10 km east): precise flags it, approximate at resolution 7 does not.
+
+    This demonstrates that H3 cell boundaries differ from exact geometry boundaries — a point outside
+    the polygon may still share H3 cells with it at coarser resolutions.
+    """
+    point = "POINT(5.20 52.35)"
+    test_df = spark.createDataFrame([[point], [None]], _GEO_SCHEMA)
+
+    precise_condition = is_geo_covers(
+        "geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True
+    )
+    actual_precise = test_df.select("geom", precise_condition)
+    expected_precise = spark.createDataFrame([[point, _precise_violation(point)], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual_precise, expected_precise, checkRowOrder=False)
+
+    approximate_condition = is_geo_covers("geom", _REF_POLYGON, resolution=5)
+    actual_approximate = test_df.select("geom", approximate_condition)
+    expected_approximate = spark.createDataFrame([[point, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual_approximate, expected_approximate, checkRowOrder=False)
+
+
+def test_is_geo_intersects_precise_interior_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point intersects the polygon — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects(
+        "geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True
+    )
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_precise_boundary_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Boundary point intersects the polygon (shared boundary point counts) — no violation."""
+    test_df = spark.createDataFrame([[_POINT_EDGE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects(
+        "geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True
+    )
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_EDGE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_precise_exterior_point_wkt_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point does not intersect the polygon — violation (WKT input)."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects(
+        "geom", _REF_POLYGON, precise=True, convert_column=True, convert_reference_geometry=True
+    )
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _precise_violation(_POINT_OUTSIDE)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_interior_point_resolution_5_no_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Interior point shares H3 cells with the polygon at resolution 5 — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_interior_point_resolution_7_no_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Interior point shares H3 cells with the polygon at resolution 7 — no violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_INSIDE, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_exterior_point_resolution_5_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point shares no H3 cells with the polygon at resolution 5 — violation."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _approximate_violation(_POINT_OUTSIDE)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_exterior_point_resolution_7_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point shares no H3 cells with the polygon at resolution 7 — violation."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _approximate_violation(_POINT_OUTSIDE)], [None, None]], _APPROXIMATE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_overlapping_polygon_resolution_5_no_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Partially overlapping polygon shares at least one H3 cell at resolution 5 — no violation."""
+    test_df = spark.createDataFrame([[_OVERLAPPING_POLYGON], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=5)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_OVERLAPPING_POLYGON, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_intersects_approximate_overlapping_polygon_resolution_7_no_violation(
+    skip_if_runtime_not_geo_compatible, spark
+):
+    """Partially overlapping polygon shares at least one H3 cell at resolution 7 — no violation."""
+    test_df = spark.createDataFrame([[_OVERLAPPING_POLYGON], [None]], _GEO_SCHEMA)
+    condition = is_geo_intersects("geom", _REF_POLYGON, resolution=7)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_OVERLAPPING_POLYGON, None], [None, None]], _APPROXIMATE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_touches_interior_point_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Interior point does not touch the polygon (touches requires shared boundary, not interior) — violation."""
+    test_df = spark.createDataFrame([[_POINT_INSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_touches("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_INSIDE, _precise_violation(_POINT_INSIDE)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_touches_boundary_point_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Boundary point touches the polygon (shared boundary, disjoint interiors) — no violation."""
+    test_df = spark.createDataFrame([[_POINT_EDGE], [None]], _GEO_SCHEMA)
+    condition = is_geo_touches("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[_POINT_EDGE, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_touches_exterior_point_wkt_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Exterior point does not touch the polygon — violation (WKT input)."""
+    test_df = spark.createDataFrame([[_POINT_OUTSIDE], [None]], _GEO_SCHEMA)
+    condition = is_geo_touches("geom", _REF_POLYGON, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[_POINT_OUTSIDE, _precise_violation(_POINT_OUTSIDE)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_within_interior_reference_no_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Reference point is inside the column polygon — st_within(reference, column) is true — no violation."""
+    column_polygon = _REF_POLYGON
+    reference_point = _POINT_INSIDE
+    test_df = spark.createDataFrame([[column_polygon], [None]], _GEO_SCHEMA)
+    condition = is_geo_within("geom", reference_point, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame([[column_polygon, None], [None, None]], _PRECISE_SCHEMA)
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_within_boundary_reference_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Reference point is on the polygon boundary — st_within excludes boundary — violation."""
+    column_polygon = _REF_POLYGON
+    reference_point = _POINT_EDGE
+    test_df = spark.createDataFrame([[column_polygon], [None]], _GEO_SCHEMA)
+    condition = is_geo_within("geom", reference_point, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[column_polygon, _precise_violation(column_polygon)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
+
+
+def test_is_geo_within_exterior_reference_violation(skip_if_runtime_not_geo_compatible, spark):
+    """Reference point is outside the column polygon — st_within is false — violation."""
+    column_polygon = _REF_POLYGON
+    reference_point = _POINT_OUTSIDE
+    test_df = spark.createDataFrame([[column_polygon], [None]], _GEO_SCHEMA)
+    condition = is_geo_within("geom", reference_point, convert_column=True, convert_reference_geometry=True)
+    actual = test_df.select("geom", condition)
+    expected = spark.createDataFrame(
+        [[column_polygon, _precise_violation(column_polygon)], [None, None]], _PRECISE_SCHEMA
+    )
+    assertDataFrameEqual(actual, expected, checkRowOrder=False)
